moved store to own file and filtered ranks above 50
Needed to filter ranks above 50, because strava returned some wierd ranks related to the provided client user
